refactor
- deps upd
- less deps now
- demo renamed to example
- webpack replaced to rspack
- ci: node ver updated to 20
