+++ .github/workflows/npm-publish.yml
@@ -1,32 +1,22 @@
-name: Publish NPM package
+name: Publish Node.js Package
 
 on:
   release:
     types: [created]
 
 jobs:
-  publish-npm:
+  build:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-
-      - uses: actions/setup-node@v2
+      - uses: actions/checkout@v4
+      - uses: actions/setup-node@v4
         with:
-          node-version: 14
+          node-version: 20
           registry-url: https://registry.npmjs.org/
-
-      - name: Install dependencies
-        run: |
-          npm ci
-          npm i -g npm@7 --registry=https://registry.npmjs.org
-
-      - name: Build package
-        run: |
-          npm --no-git-tag-version version ${GITHUB_REF#refs/*/}
-          npm set-script prepare ""
-          npm run build
-
-      - name: Publish build
-        run: npm publish --access public
+      - run: npm ci
+      - run: npm run check
+      - run: npm run build
+      - run: npm --no-git-tag-version version ${{ github.event.release.tag_name }}
+      - run: npm publish --access public
         env:
-          NODE_AUTH_TOKEN: ${{secrets.npm_token}}
+          NODE_AUTH_TOKEN: ${{ secrets.npm_token }}
