--- .npmignore
@@ -1,5 +1,0 @@
-/dist
-/build
-/node_modules
-.DS_Store
-.cache
