--- demo/.swcrc
@@ -1,14 +1,0 @@
-{
-  "jsc": {
-    "parser": {
-      "syntax": "typescript",
-      "tsx": true
-    },
-    "target": "es2020",
-    "transform": {
-      "react": {
-        "runtime": "automatic"
-      }
-    }
-  }
-}
