--- demo/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "rglk-demo-page",
-  "private": true,
-  "version": "0.0.0",
-  "description": "Demo page of rglk library",
-  "scripts": {
-    "start": "webpack serve --mode=development",
-    "build": "webpack --mode=production",
-    "type-check": "tsc -p . --noEmit"
-  },
-  "author": "Dmitry Petrov",
-  "license": "Apache-2.0",
-  "devDependencies": {
-    "@swc/core": "^1.2.171",
-    "css-loader": "^6.7.1",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.6.0",
-    "swc-loader": "^0.2.0",
-    "typescript": "^4.6.4",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.8.1"
-  },
-  "dependencies": {
-    "@types/react": "^18.0.8",
-    "@types/react-dom": "^18.0.3",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0"
-  }
-}
