--- demo/src/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>rglk.js demo</title>
-  </head>
-  <body>
-    <div id="main"></div>
-  </body>
-</html>
