--- demo/webpack.config.js
@@ -1,61 +1,0 @@
-const path = require('path');
-const HtmlPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const CssMinimizerPlugin = require('css-minimizer-webpack-plugin');
-
-const config = {
-  entry: './src/index.tsx',
-  output: {
-    path: path.resolve(__dirname, 'dist'),
-    filename: 'index.js',
-  },
-  resolve: {
-    extensions: ['.js', '.jsx', '.ts', '.tsx'],
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx|ts|tsx)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'swc-loader',
-        },
-      },
-
-      // css
-      {
-        test: /\.css$/i,
-        exclude: /\.module\.css$/i,
-        use: [MiniCssExtractPlugin.loader, 'css-loader'],
-      },
-
-      // css modules
-      {
-        test: /\.module\.css$/i,
-        use: [
-          MiniCssExtractPlugin.loader,
-          {
-            loader: 'css-loader',
-            options: {
-              modules: true,
-            },
-          },
-        ],
-      },
-    ],
-  },
-  plugins: [
-    new MiniCssExtractPlugin({
-      filename: 'index.css',
-    }),
-    new HtmlPlugin({
-      publicPath: './',
-      template: path.resolve(__dirname, 'src/index.html'),
-    }),
-  ],
-  optimization: {
-    minimizer: ['...', new CssMinimizerPlugin()],
-  },
-};
-
-module.exports = config;
