+++ example/src/components/section/index.tsx
