+++ example/src/typings/css-modules.d.ts
@@ -2,3 +2,8 @@ declare module '*.module.css' {
   const classes: { [key: string]: string };
   export default classes;
 }
+
+declare module '*.m.css' {
+  const classes: { [key: string]: string };
+  export default classes;
+}
