--- jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.(t|j)sx?$': ['@swc/jest'],
-  },
-};
