--- lint-staged.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  '*.{js,jsx,ts,tsx}': ['prettier --write'],
-  '*.{css,scss}': ['prettier --write'],
-};
