+++ package.json
@@ -1,19 +1,13 @@
 {
   "name": "rglk",
+  "type": "module",
   "version": "0.0.0",
   "license": "MIT",
   "description": "Simple JS library for development roguelike games",
   "author": "Dmitry Petrov",
-  "files": [
-    "dist",
-    "src",
-    "readme.md"
-  ],
-  "main": "./dist/index.js",
   "keywords": [
     "roguelike",
     "game",
-    "development",
     "dungeon",
     "astar",
     "fov"
@@ -23,20 +17,24 @@
     "url": "https://github.com/krutoo/rglk.js.git"
   },
   "scripts": {
-    "test": "jest",
-    "build:before": "rm -rf dist",
+    "check": "tsc -p . --noEmit",
+    "fmt": "prettier . --write",
+    "test": "node --import=tsimp/import --test src/prng/__test__/index.test.ts",
+    "build:clean": "rm -rf dist",
     "build:main": "tsc --project tsconfig.build.json",
-    "build": "npm run build:before && npm run build:main",
-    "type-check": "tsc -p . --noEmit",
-    "prepare": "husky install"
+    "build": "npm run build:clean && npm run build:main"
   },
   "devDependencies": {
-    "@swc/jest": "^0.2.20",
-    "@types/jest": "^27.4.1",
-    "husky": "^7.0.4",
-    "jest": "^28.0.3",
-    "lint-staged": "^12.4.1",
-    "prettier": "^2.6.2",
-    "typescript": "^4.6.4"
+    "@types/node": "^22.8.7",
+    "prettier": "^3.3.3",
+    "tsimp": "^2.0.12",
+    "typescript": "^5.6.3"
+  },
+  "files": [
+    "dist",
+    "README.md"
+  ],
+  "exports": {
+    ".": "./dist/index.js"
   }
 }
