+++ src/prng/__test__/index.test.ts
@@ -1,19 +1,11 @@
-import { createGenerator } from '..';
+import { createPRNG } from '../index.js';
+import { test, describe } from 'node:test';
+import assert from 'node:assert';
 
 describe('createGenerator', () => {
-  test('should throw when argument is not a finite number', () => {
-    const invalidValues = ['1', true, null, undefined, {}, Object, Symbol('test')];
-
-    invalidValues.forEach(value => {
-      expect(() => createGenerator(value as any)).toThrow(
-        'First argument "seed" must be a finite number.',
-      );
-    });
-  });
-
   test('should number generators that based on seed', () => {
-    const firstGen = createGenerator(123);
-    const secondGen = createGenerator(123);
+    const firstGen = createPRNG(123);
+    const secondGen = createPRNG(123);
 
     let firstGenValue;
     let secondGenValue;
@@ -21,8 +13,8 @@ describe('createGenerator', () => {
     for (let i = 0; i < 10; i++) {
       firstGenValue = firstGen();
       secondGenValue = secondGen();
-      expect(typeof firstGenValue).toBe('number');
-      expect(firstGenValue).toBe(secondGenValue);
+      assert.strictEqual(typeof firstGenValue, 'number');
+      assert.strictEqual(firstGenValue, secondGenValue);
     }
   });
 });
